Fix Long→Int overflow in chart data mapping (#22778)
* Remove unnecessary Long→Int conversion in chart data mapping

Vico's series() accepts Number, so Long values can be passed
directly without risking silent overflow for counts above 2.1B.


* Remove unnecessary Long→Int conversion in chart data mapping

Vico's series() accepts Number, so Long values can be passed
directly without risking silent overflow for counts above 2.1B.


* Update wordpress-rs library version (CMM-1990)


* Revert "Update wordpress-rs library version (CMM-1990)"

This reverts commit d5e49f407e2b5dfb64967c9667451d53fcc57a41.

---------

+++ WordPress/src/main/java/org/wordpress/android/ui/newstats/todaysstats/TodaysStatsCard.kt
@@ -352,10 +352,10 @@ private fun StatsChart(chartData: ChartData) {
             modelProducer.runTransaction {
                 lineSeries {
                     // Today's data (solid line)
-                    series(chartData.currentPeriod.map { it.views.toInt() })
+                    series(chartData.currentPeriod.map { it.views })
                     // Yesterday's data (dashed line) - only if available
                     if (hasPreviousPeriod) {
-                        series(chartData.previousPeriod.map { it.views.toInt() })
+                        series(chartData.previousPeriod.map { it.views })
                     }
                 }
             }
